+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #249 Remove APIs deprecated since 0.16
 * #250 Remove APIs deprecated since 0.17 
+* #251 Remove APIs deprecated since 0.18
 
 ## Release 1.0.0-M19